Move xssFilter into a seperate file out of documentFormatter
