+++ sashimi-webapp/src/logic/documentPackager/index.js
@@ -1,5 +1,6 @@
 import markdownProcessor from './markdownProcessor';
 import documentFormatter from './documentFormatter';
+import xssFilter from './xssFilter';
 
 const processData = function processData(data) {
   return markdownProcessor.process(data);
@@ -11,6 +12,7 @@ const formatData = function formatData(data) {
 
 export default {
   render: function render(data) {
-    return formatData(processData(data));
+    const formattedData = formatData(processData(data));
+    return xssFilter.filter(formattedData);
   }
 };
